Big data
In this, there was a huge amount of data having unsorted data , here in this Dataset we created a DataFrame using Pandas Library and did the data cleaning and Dropped few unnecessary columns and made the dataset fit for visualization. This .CSV file has 23k Data that we reduced to achieve the visualization , there were also some Male and females having age 110 years 114 years 116 years which was the outliers hence we removed those data as well.